SB20-465: Al cancelar pedido se recargan valores de pedidos (totales, cantidad, etc) en la pagina. Se hicieron algunos ajustes:
- El método UsuarioService.GetMisPedidosDetalleConsultoraOnline paso el parámetro PedidoID de int a long
- Se agregaron enum y contanstes para manejar tipo de atención y mensajes de cancelacion exitosa.
- Se agregaron los métodos de ConsultoraOnline en MisPedidosController a Global.asax, para validar sesión.

+++ app/Portal.Consultoras.Service/UsuarioService.cs
@@ -308,7 +308,7 @@ namespace Portal.Consultoras.Service
             return BLMisPedidos.GetMisPedidos(PaisID, ConsultoraId, Campania);
         }
 
-        public IList<BEMisPedidosDetalle> GetMisPedidosDetalleConsultoraOnline(int PaisID, int PedidoID)
+        public IList<BEMisPedidosDetalle> GetMisPedidosDetalleConsultoraOnline(int PaisID, long PedidoID)
         {
             var BLMisPedidos = new BLConsultoraOnline();
             return BLMisPedidos.GetMisPedidosDetalle(PaisID, PedidoID);
